requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.1
celery[redis]==5.4.0
redis==5.2.1
pydantic[email]==2.10.4
pydantic-settings==2.7.1
python-jose[cryptography]==3.3.0
cryptography==44.0.0
loguru==0.7.3
httpx==0.28.1
beautifulsoup4==4.12.3
sentry-sdk[fastapi]==2.19.2
faker==33.1.0
psycopg2-binary